requirements.txt
about-time==3.1.1
alive-progress==2.4.1
appnope==0.1.3
asttokens==2.1.0
backcall==0.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
decorator==5.1.1
executing==1.2.0
fonttools==4.38.0
grapheme==0.6.0
idna==3.4
ipython==8.6.0
jedi==0.18.2
joblib==1.2.0
kafka-python3==3.0.0
kiwisolver==1.4.4
matplotlib==3.6.2
matplotlib-inline==0.1.6
nltk==3.7
numpy==1.23.4
oauthlib==3.2.2
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
prompt-toolkit==3.0.33
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
stack-data==0.6.1
tqdm==4.64.1
traitlets==5.5.0
tweepy==4.12.1
urllib3==1.26.12
wcwidth==0.2.5
wordcloud==1.8.2.2
